requirements.txt
asgiref==3.8.1
Django==5.1.4
sqlparse==0.5.3
tzdata==2024.2
djangorestframework==3.15.2